+++ frontend/themes/gedu/site/common/elegant.php
@@ -57,4 +57,4 @@
                     <div class="clearfix"></div>
                 </div>
             </div> 
-        </div>
+
