+++ frontend/web/css/gstyle.css
@@ -428,7 +428,9 @@
   	color: white;
   	margin-left: 5px;
   }
-
+.demo .teabor .teacherdetail{
+	margin-top: 20px;
+}
 
  .row-box .headcolor{color: black;}
   @media only screen and (max-width:990px){
